+++ Rebus/Persistence/SqlServer/SqlServerSagaStorage.cs
@@ -63,7 +63,12 @@ namespace Rebus.Persistence.SqlServer
 
                 var hasDataTable = tableNames.Contains(_dataTableName, StringComparer.OrdinalIgnoreCase);
                 var hasIndexTable = tableNames.Contains(_indexTableName, StringComparer.OrdinalIgnoreCase);
-                
+
+                if (hasDataTable)
+                {
+                    VerifyDataTableSchema(_dataTableName, connection);
+                }
+
                 if (hasDataTable && hasIndexTable)
                 {
                     return;
@@ -85,30 +90,30 @@ namespace Rebus.Persistence.SqlServer
 
                 using (var command = connection.CreateCommand())
                 {
-                    command.CommandText = string.Format(@"
-CREATE TABLE [dbo].[{0}] (
+                    command.CommandText = $@"
+CREATE TABLE [dbo].[{_dataTableName}] (
 	[id] [uniqueidentifier] NOT NULL,
 	[revision] [int] NOT NULL,
 	[data] [varbinary](max) NOT NULL,
-    CONSTRAINT [PK_{0}] PRIMARY KEY CLUSTERED 
+    CONSTRAINT [PK_{_dataTableName}] PRIMARY KEY CLUSTERED 
     (
 	    [id] ASC
     )
 )
-", _dataTableName);
+";
 
                     await command.ExecuteNonQueryAsync();
                 }
 
                 using (var command = connection.CreateCommand())
                 {
-                    command.CommandText = string.Format(@"
-CREATE TABLE [dbo].[{0}] (
+                    command.CommandText = $@"
+CREATE TABLE [dbo].[{_indexTableName}] (
 	[saga_type] [nvarchar](40) NOT NULL,
 	[key] [nvarchar](200) NOT NULL,
 	[value] [nvarchar](200) NOT NULL,
 	[saga_id] [uniqueidentifier] NOT NULL,
-    CONSTRAINT [PK_{0}] PRIMARY KEY CLUSTERED 
+    CONSTRAINT [PK_{_indexTableName}] PRIMARY KEY CLUSTERED 
     (
 	    [key] ASC,
 	    [value] ASC,
@@ -116,23 +121,24 @@ CREATE TABLE [dbo].[{0}] (
     )
 )
 
-CREATE NONCLUSTERED INDEX [IX_{0}_saga_id] ON [dbo].[{0}]
+CREATE NONCLUSTERED INDEX [IX_{_indexTableName}_saga_id] ON [dbo].[{_indexTableName}]
 (
 	[saga_id] ASC
 )
-", _indexTableName);
+";
 
                     await command.ExecuteNonQueryAsync();
                 }
 
                 using (var command = connection.CreateCommand())
                 {
-                    command.CommandText = string.Format(@"
-ALTER TABLE [dbo].[{0}] WITH CHECK 
-    ADD CONSTRAINT [FK_{1}_id] FOREIGN KEY([saga_id])
+                    command.CommandText =
+                        $@"
+ALTER TABLE [dbo].[{_indexTableName}] WITH CHECK 
+    ADD CONSTRAINT [FK_{_dataTableName}_id] FOREIGN KEY([saga_id])
 
-REFERENCES [dbo].[{1}] ([id]) ON DELETE CASCADE
-", _indexTableName, _dataTableName);
+REFERENCES [dbo].[{_dataTableName}] ([id]) ON DELETE CASCADE
+";
 
                     await command.ExecuteNonQueryAsync();
                 }
@@ -148,6 +154,43 @@ ALTER TABLE [dbo].[{_indexTableName}] CHECK CONSTRAINT [FK_{_dataTableName}_id]
                 }
 
                 await connection.Complete();
+            }
+        }
+
+        void VerifyDataTableSchema(string dataTableName, IDbConnection connection)
+        {
+            //  [id] [uniqueidentifier] NOT NULL,
+            //	[revision] [int] NOT NULL,
+            //	[data] [varbinary](max) NOT NULL,
+            var expectedDataTypes = new Dictionary<string, SqlDbType>(StringComparer.InvariantCultureIgnoreCase)
+            {
+                {"id", SqlDbType.UniqueIdentifier },
+                {"revision", SqlDbType.Int },
+                {"data", SqlDbType.VarBinary },
+            };
+
+            var columns = connection.GetColumns(dataTableName);
+
+            foreach (var column in columns)
+            {
+                // we skip columns we don't know about - don't prevent people from adding their own columns
+                if (!expectedDataTypes.ContainsKey(column.Name)) continue;
+
+                var expectedDataType = expectedDataTypes[column.Name];
+
+                if (column.Type == expectedDataType) continue;
+
+                // special case: migrating from Rebus 0.99.59 to 0.99.60
+                if (column.Name == "data" && column.Type == SqlDbType.NVarChar && expectedDataType == SqlDbType.VarBinary)
+                {
+                    throw new RebusApplicationException(@"Sorry, but the [data] column data type was changed from NVarChar(MAX) to VarBinary(MAX) in Rebus 0.99.60.
+
+This was done because it turned out that SQL Server was EXTREMELY SLOW to load a saga's data when it was saved as NVarChar - you can expect a reduction in saga data loading time to about 1/10 of the previous time from Rebus version 0.99.60 and on.
+
+Unfortunately, Rebus cannot help migrating any existing pieces of saga data :( so we suggest you wait for a good time when the saga data table is empty, and then you simply wipe the tables and let Rebus (re-)create them.");
+                }
+
+                throw new RebusApplicationException($"The column [{column.Name}] has the type {column.Type} and not the expected {expectedDataType} data type!");
             }
         }