+++ CHANGELOG.md
@@ -959,7 +959,7 @@
 
 # 0.99.60
 
-* Changed data type of the SQL Server saga storage `[data]` column to `VARBINARY(MAX)` because `NVARCHAR(MAX)` is extremely slow!
+* Changed data type of the SQL Server saga storage `[data]` column to `VARBINARY(MAX)` because `NVARCHAR(MAX)` was extremely slow!
 
 ---